+++ lib/spior/persist.rb
@@ -2,6 +2,7 @@ require 'nomansland'
 require 'tty-which'
 require_relative 'copy'
 require_relative 'msg'
+require_relative 'helpers'
 
 module Spior
   module Persist
@@ -17,10 +18,15 @@ module Spior
 
     def search_for_systemd
       return if !TTY::Which.exist?('systemctl') 
+      @systemctl = Helpers::Exec.new("systemctl")
+      @iptables_save = Helpers::Exec.new("iptables-save")
       Spior::Copy::systemd_services
       @services.each do |service|
         Msg.p "Search for service #{service}..."
-        system("if ! systemctl is-enabled #{service} ; then sudo systemctl enable #{service} ; fi")
+        `systemctl is-enabled #{service}`
+        if not $?.success? then
+          @systemctl.run("enable #{service}")
+        end
       end
       iptables_systemd
     end
@@ -28,9 +34,11 @@ module Spior
     def iptables_systemd
       case Nomansland::installer?
       when :pacman
-        system('sudo iptables-save -f /etc/iptables/iptables.rules')
+        @iptables_save.run("-f /etc/iptables/iptables.rules")
       when :emerge
-        system('sudo systemctl start iptables-store')
+        @systemctl.run("start iptables-store")
+      when :apt_get
+        @iptables_save.run("> /etc/iptables/rules.v4")
       else
         Msg.report "Fail for save iptables-rule, your system is not yet supported"
       end